// src/app/model/md-name.ts
export enum MdName {
    //share market related constants

    SHARE_MARKET = 'Share Market',
    PUBLIC_COMPANY = 'Public Company',
    SHARE = 'Share',
    STOCK = 'Stock',
    NSE = 'NSE',
    BSE = 'BSE',
    INDEXES = 'Indexes',
    INDEXES_FORMATION = 'Indexes Formation',

    //broker related constants
    BROKER = 'Broker',
    BROKERS_IN_INDIA = 'Brokers In India',
    BROKERAGE_CHARGES = 'Brokerage Charges',

    //mutual fund related constants
    MUTUAL_FUND = 'Mutual Fund',
    MUTUAL_FUND_TYPES = 'Mutual Fund Types',
    INDEX_MUTUAL_FUND = 'Index Mutual Fund',
    MUTUAL_FUND_PROVIDER = 'Mutual Fund Provider',
    MUTUAL_FUND_CHARGES = 'Mutual Fund Charges',

    //tax related constants
    TAX_ON_STOCK = 'Tax On Stock',
    TAX_ON_MUTUAL_FUND = 'Tax On Mutual Fund',

    //governing related constants
    SEBI = 'SEBI',
    NSDL = 'NSDL',
    CDSL = 'CDSL',

    //investment related constants
    TYPES_OF_INVESTMENT = 'Types Of Investment'

}